+++ t/test-trs-recover2.sh
@@ -17,7 +17,7 @@
 # Check parallel harness features:
 #  - recovery from unreadable '.trs' files, in various scenarios
 
-. ./defs || Exit 1
+. ./defs || exit 1
 
 cat >> configure.ac <<END
 AC_OUTPUT
@@ -41,7 +41,7 @@ $AUTOMAKE -a
 
 : > t
 chmod a-r t
-test ! -r t || Exit 77
+test ! -r t || exit 77
 rm -f t
 
 : Create the required log files.
@@ -70,7 +70,7 @@ test -r bar.trs
 : Again, but using "make recheck" this time.
 rm -f foo.trs
 chmod a-r bar.trs
-$MAKE recheck >stdout || { cat stdout; Exit 1; }
+$MAKE recheck >stdout || { cat stdout; exit 1; }
 cat stdout
 test -f foo.trs
 test -r foo.trs
@@ -82,12 +82,12 @@ grep '^PASS: bar\.test' stdout
 : Recreate by remaking the global test log.
 chmod a-r foo.trs
 rm -f test-suite.log
-$MAKE test-suite.log >stdout || { cat stdout; Exit 1; }
+$MAKE test-suite.log >stdout || { cat stdout; exit 1; }
 cat stdout
 test -f foo.trs
 test -r foo.trs
 grep '^PASS: foo\.test' stdout
-grep 'bar\.test' stdout && Exit 1
+grep 'bar\.test' stdout && exit 1
 # Also test that have only run before should be counted in the final
 # testsuite summary.
 grep '^# TOTAL:  *2$' stdout
@@ -95,7 +95,7 @@ grep '^# TOTAL:  *2$' stdout
 : Setup for the next check.
 : > baz.test
 sed 's/^TESTS =.*/& baz.test/' Makefile > t
-diff t Makefile && Exit 99
+diff t Makefile && exit 99
 mv -f t Makefile
 $MAKE check
 test -f foo.trs
@@ -108,13 +108,13 @@ $sleep
 touch stamp
 $sleep
 touch bar.test
-RECHECK_LOGS= $MAKE -e check >stdout || { cat stdout; Exit 1; }
+RECHECK_LOGS= $MAKE -e check >stdout || { cat stdout; exit 1; }
 cat stdout
 test -r foo.trs
 is_newest bar.trs bar.test
 grep '^PASS: foo\.test' stdout
 grep '^PASS: bar\.test' stdout
-grep 'baz\.test' stdout && Exit 1
+grep 'baz\.test' stdout && exit 1
 # Also test that have only run before should be counted in the final
 # testsuite summary.
 grep '^# TOTAL:  *3$' stdout
